feat(085): split contracts by context
